Images should display on the main screen now. Other work in tests

+++ app/src/main/java/com/davidgrath/expensetracker/entities/db/ImageDb.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class ImageDb(
     @PrimaryKey(autoGenerate = true)
     val id: Long?,
-    val sizeBytes: Int,
+    val sizeBytes: Long,
     val sha256: String,
     val mimeType: String,
     val uri: String,
